feat(api_token) improve migration tests
